==> scripts/09_map_clusters.ipynb <==
import pandas as pd
from collections import Counter


df = pd.read_pickle("../results/pre_final_files/schembl_summs_v3_alg_cleaned_labels.pkl")

vocab_df = pd.read_csv("../results/pre_final_files/schembl_summs_v4_vocab_gpt_cleaned_eps_0.340_diff_20030_clusters_errors_fixed.csv")

vocab_df.loc[0, "gpt_cleaned_labels"] = ""

# remove row with API REQUEST ERROR
# vocab_df = vocab_df[vocab_df["gpt_cleaned_labels"] != "API REQUEST ERROR"]

print(len(vocab_df))

words = set()
for i in vocab_df['gpt_cleaned_labels']:
    words.update([i])
print(len(words))

# print clusters with more than 2 words (when gpt messes up / cannot determine)
print(vocab_df[vocab_df['gpt_cleaned_labels'].str.split().str.len() > 2]["gpt_cleaned_labels"])

# drop clusters with more than 2 words (when gpt messes up / cannot determine)
vocab_df = vocab_df[~(vocab_df['gpt_cleaned_labels'].str.split().str.len() > 2)].reset_index(drop=True)

words = set()
for i in vocab_df['gpt_cleaned_labels']:
    words.update([i])
print(len(words))

# make each element in original_clustered_labels its own row, repeating the gpt_cleaned_labels
vocab_df = vocab_df.assign(original_clustered_labels=vocab_df['original_clustered_labels'].str.split(',')).explode('original_clustered_labels')
vocab_df = vocab_df.reset_index()
# convert to dictionary
vocab_dict = dict(zip(vocab_df["original_clustered_labels"].tolist(), vocab_df["gpt_cleaned_labels"].tolist()))

# remove leading spaces from keys
vocab_dict = {k.strip(): v for k, v in vocab_dict.items()}

df["summarizations"] = df["summarizations"].map(set)

# map vocab_dict to each value in summarizations, ignoring when the value is not in vocab_dict and defaulting to the original value
df['summarizations'] = df['summarizations'].map(lambda x: [vocab_dict.get(i, i) for i in x])

# remove '&' label
df['summarizations'] = df['summarizations'].map(lambda x: [i for i in x if i != '&'])

words = set()
for i in df["summarizations"]:
    
    words.update(set(i))
print(len(words))
# remove empty strings from df["summarizations"]
df["summarizations"] = df["summarizations"].apply(lambda x: [i for i in x if i != ""])

words = set()
for i in df["summarizations"]:
    words.update(set(i))
print(len(words))

# count all words in df["summarizations"]
counter = Counter()
for summarization in df["summarizations"]:
    counter.update(summarization)
counter.most_common()

df[["smiles", "cid", "patent_ids", "summarizations"]].to_csv("../results/schembl_summs_v4_gpt_cleaned_eps_0.340_diff_20030_final.csv")

# remove all entries if they contain less than 10 counts

counter_trimmed = counter.copy()
for key in counter.keys():
    if counter[key] < 50:
        del counter_trimmed[key]

print(len(counter_trimmed))

# remove entries from df["summariations"] if they are not in counter_trimmed
df["summarizations"] = df["summarizations"].apply(lambda x: [y for y in x if y in counter_trimmed])

# save counter labels in order of count

with open('../results/all_labels.txt', 'w') as f:
    for label, count in counter_trimmed.most_common():
        f.write("%s\n" % label)

print(len(df))

df = df.dropna(subset=["summarizations"])

print(len(df))

df[["smiles", "cid", "patent_ids", "summarizations"]].to_csv("../results/schembl_summs_v5_final.csv", index=False)

from rdkit import Chem
df["fingerprint"] = df["smiles"].apply(lambda x: Chem.RDKFingerprint(Chem.MolFromSmiles(x)))

import numpy as np
df["fingerprint"] = df["fingerprint"].apply(np.array)

df[["smiles", "cid", "patent_ids", "summarizations", "fingerprint"]].to_pickle("../results/schembl_summs_v5_final_fp.pkl")
